requirements.txt
torch==2.1.0
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
opencv-python
pillow
fastapi
uvicorn
